Break out LogMachine into separate file, and use XSBUG_LOGMACHINE env. variable to conditionalize the implementation to load
