+++ tools/xsbug-log/xsbug-log.js
@@ -18,12 +18,15 @@
  *
  */
 
-let net, exec, machine;
+let net, exec, LogMachine;
 
 try {
 	net = require('node:net');
 	exec = require('node:child_process').exec;
-	Machine = require('./xsbug-machine.js').Machine;
+	if (process.env.XSBUG_LOGMACHINE)
+		LogMachine = require(process.env.XSBUG_LOGMACHINE).LogMachine;
+	else
+		LogMachine = require('./xsbug-logmachine.js').LogMachine;
 }
 catch (e) {
 	if ("MODULE_NOT_FOUND" === e.code) {
@@ -34,40 +37,6 @@ catch (e) {
 	else
 		console.log("#xsbug-log start-up error: " + e);
 	process.exit();
-}
-
-class LogMachine extends Machine {
-	view = {};
-	log = "";
-
-	onTitleChanged(title, tag) {
-		super.onTitleChanged(title, tag);
-		
-		if (title && (title !== "mcsim"))
-			console.log(`#xsbug-log connected to "${title}"`);
-
-		this.doSetAllBreakpoint([], false, true);		// break on exceptions
-	}
-	onLogged(path, line, data) {
-		this.log += data;
-		if (this.log.endsWith("\n")) {
-			console.log(this.log.slice(0, this.log.length - 1));
-			this.log = "";
-		}
-	}
-	onBroken(path, line, text) {
-		this.view.frames.forEach((frame, index) => {
-			let line = "  #" + index + ": " + frame.name;
-			if (frame.path)
-				line += " " + frame.path + ":" + frame.line
-			console.log(line);
-		});
-
-		super.onBroken(path, line, text);
-	}
-	onViewChanged(name, items) {
-		this.view[name] = items;
-	}
 }
 
 const portIn = 5002;
