+++ tools/xsbug-log/xsbug-logmachine.js
@@ -1,0 +1,37 @@
+let Machine = require('./xsbug-machine.js').Machine;
+
+class LogMachine extends Machine {
+	view = {};
+	log = "";
+
+	onTitleChanged(title, tag) {
+		super.onTitleChanged(title, tag);
+		
+		if (title && (title !== "mcsim"))
+			console.log(`#xsbug-log connected to "${title}"`);
+
+		this.doSetAllBreakpoint([], false, true);		// break on exceptions
+	}
+	onLogged(path, line, data) {
+		this.log += data;
+		if (this.log.endsWith("\n")) {
+			console.log(this.log.slice(0, this.log.length - 1));
+			this.log = "";
+		}
+	}
+	onBroken(path, line, text) {
+		this.view.frames.forEach((frame, index) => {
+			let line = "  #" + index + ": " + frame.name;
+			if (frame.path)
+				line += " " + frame.path + ":" + frame.line
+			console.log(line);
+		});
+
+		super.onBroken(path, line, text);
+	}
+	onViewChanged(name, items) {
+		this.view[name] = items;
+	}
+}
+
+exports.LogMachine = LogMachine;
